feat: nick unique check
