+++ others/bircf/.vscode/targets.log
@@ -7,16 +7,15 @@ make all --print-data-base --no-builtin-variables --no-builtin-rules --question
 
 # This program built for i386-apple-darwin11.3.0
  
-make: *** No rule to make target `main.o', needed by `bircf'.  Stop.
 
-
-# Make data base, printed on Thu Jun  9 16:54:09 2022
+# Make data base, printed on Thu Jun  9 17:34:00 2022
 
 # Variables
 
 # automatic
 <D = $(patsubst %/,%,$(dir $<))
 # automatic
+ 
 ?F = $(notdir $?)
 # environment
 ELECTRON_NO_ATTACH_CONSOLE = 1
@@ -29,11 +28,10 @@ VSCODE_LOG_NATIVE = false
 # automatic
 @F = $(notdir $@)
 # makefile
- 
 CURDIR := /Users/kyuhokim/Cadet/chirc/others/bircf
 # makefile
 SHELL = /bin/sh
-# makefile (from `Makefile', line 14)
+# makefile (from `Makefile', line 15)
 RM = rm -f
 # environment
 VSCODE_NLS_CONFIG = {"locale":"ko","availableLanguages":{},"_languagePackSupport":true}
@@ -41,7 +39,7 @@ VSCODE_NLS_CONFIG = {"locale":"ko","availableLanguages":{},"_languagePackSupport
 _ = /usr/bin/make
 # makefile (from `Makefile', line 1)
 MAKEFILE_LIST :=  Makefile
-# makefile (from `Makefile', line 10)
+# makefile (from `Makefile', line 11)
 CXXFLAGS = -I. -g3 -Wall -Wextra -Werror 
 # environment
 VSCODE_VERBOSE_LOGGING = true
@@ -61,7 +59,7 @@ LSCOLORS = Gxfxcxdxbxegedabagacad
 VSCODE_LOG_STACK = false
 # environment
 ELECTRON_RUN_AS_NODE = 1
-# makefile (from `Makefile', line 11)
+# makefile (from `Makefile', line 12)
 LDFLAGS = 
 # default
 .FEATURES := target-specific order-only second-expansion else-if archives jobserver check-symlink
@@ -71,7 +69,7 @@ SSH_AUTH_SOCK = /private/tmp/com.apple.launchd.VNAmwG83DD/Listeners
 %F = $(notdir $%)
 # environment
 VSCODE_PIPE_LOGGING = true
-# makefile (from `Makefile', line 6)
+# makefile (from `Makefile', line 7)
 OBJS = ${SRCS:.cpp=.o}
 # environment
 PWD = /Users/kyuhokim/Cadet/chirc/others/bircf
@@ -108,7 +106,7 @@ MAKE = $(MAKE_COMMAND)
 # default
 MAKECMDGOALS := all
 # makefile (from `Makefile', line 2)
-SRCS = main.cpp MessageManager.cpp Commands/Pass.cpp Commands/Nick.cpp 
+SRCS = main.cpp MessageManager.cpp Commands/Pass.cpp Commands/Nick.cpp Commands/Privmsg.cpp 
 # environment
 SHLVL = 3
 # default
@@ -119,7 +117,7 @@ USER = kyuhokim
 .DEFAULT_GOAL := bircf
 # environment
 TERM_SESSION_ID = 2070FFE1-1D4F-40DF-8059-BA2BFE93651E
-# makefile (from `Makefile', line 8)
+# makefile (from `Makefile', line 9)
 NAME = bircf
 # environment
 LESS = -R
@@ -149,6 +147,7 @@ TERM_PROGRAM_VERSION = 445
 XPC_SERVICE_NAME = application.com.microsoft.VSCode.8875144.8875150.36DE7DD8-B672-485F-BD1A-CCD0CE13F327
 # environment
 HOMEBREW_PREFIX = /opt/homebrew
+ 
 # automatic
 +D = $(patsubst %/,%,$(dir $+))
 # automatic
@@ -163,7 +162,7 @@ COMMAND_MODE = unix2003
 MAKEFILES := 
 # automatic
 <F = $(notdir $<)
-# makefile (from `Makefile', line 13)
+# makefile (from `Makefile', line 14)
 CXX = c++
 # environment
 PAGER = less
@@ -190,9 +189,9 @@ VSCODE_PID = 73028
 
 # Directories
 
-# . (device 16777233, inode 11864172): 12 files, no impossibilities.
+# . (device 16777233, inode 11864172): 15 files, no impossibilities.
 
-# 12 files, no impossibilities in 1 directories.
+# 15 files, no impossibilities in 1 directories.
 
 # Implicit Rules
 
@@ -204,7 +203,8 @@ all: bircf
 #  Command-line target.
 #  Implicit rule search has been done.
 #  File does not exist.
-#  File has not been updated.
+#  File has been updated.
+#  Successfully updated.
 # variable set hash-table stats:
 # Load=0/32=0%, Rehash=0, Collisions=0/0=0%
 
@@ -217,7 +217,7 @@ all: bircf
 # Not a target:
 Makefile:
 #  Implicit rule search has been done.
-#  Last modified 2022-06-09 16:54:08
+#  Last modified 2022-06-09 17:33:18
 #  File has been updated.
 #  Successfully updated.
 # variable set hash-table stats:
@@ -227,17 +227,18 @@ fclean: clean
 #  Implicit rule search has not been done.
 #  Modification time never checked.
 #  File has not been updated.
-#  commands to execute (from `Makefile', line 25):
+#  commands to execute (from `Makefile', line 26):
 	${RM} ${NAME}
 	
 
-bircf: main.o MessageManager.o Commands/Pass.o Commands/Nick.o
+bircf: main.o MessageManager.o Commands/Pass.o Commands/Nick.o Commands/Privmsg.o
 #  Implicit rule search has not been done.
-#  File does not exist.
-#  File has not been updated.
+#  Last modified 2022-06-09 17:33:31
+#  File has been updated.
+#  Successfully updated.
 # variable set hash-table stats:
 # Load=0/32=0%, Rehash=0, Collisions=0/0=0%
-#  commands to execute (from `Makefile', line 17):
+#  commands to execute (from `Makefile', line 18):
 	${CXX} ${CXXFLAGS} -o ${NAME} ${OBJS} ${LDFLAGS}
 	
 
@@ -249,9 +250,12 @@ bircf: main.o MessageManager.o Commands/Pass.o Commands/Nick.o
 
 # Not a target:
 Commands/Nick.o:
-#  Implicit rule search has not been done.
-#  Modification time never checked.
-#  File has not been updated.
+#  Implicit rule search has been done.
+#  Last modified 2022-06-09 17:33:31
+#  File has been updated.
+#  Successfully updated.
+# variable set hash-table stats:
+# Load=0/32=0%, Rehash=0, Collisions=0/0=0%
 
 re: fclean all
 #  Implicit rule search has not been done.
@@ -262,32 +266,48 @@ clean:
 #  Implicit rule search has not been done.
 #  Modification time never checked.
 #  File has not been updated.
-#  commands to execute (from `Makefile', line 22):
+#  commands to execute (from `Makefile', line 23):
 	${RM} ${OBJS} *~ #*#
 	
 
 # Not a target:
 main.o:
 #  Implicit rule search has been done.
-#  File does not exist.
-#  File has not been updated.
+#  Last modified 2022-06-09 17:33:29
+#  File has been updated.
+#  Successfully updated.
 # variable set hash-table stats:
 # Load=0/32=0%, Rehash=0, Collisions=0/0=0%
 
 # Not a target:
 MessageManager.o:
-#  Implicit rule search has not been done.
-#  Modification time never checked.
-#  File has not been updated.
+#  Implicit rule search has been done.
+#  Last modified 2022-06-09 17:33:30
+#  File has been updated.
+#  Successfully updated.
+# variable set hash-table stats:
+# Load=0/32=0%, Rehash=0, Collisions=0/0=0%
+
+# Not a target:
+Commands/Privmsg.o:
+#  Implicit rule search has been done.
+#  Last modified 2022-06-09 17:33:31
+#  File has been updated.
+#  Successfully updated.
+# variable set hash-table stats:
+# Load=0/32=0%, Rehash=0, Collisions=0/0=0%
 
 # Not a target:
 Commands/Pass.o:
-#  Implicit rule search has not been done.
-#  Modification time never checked.
-#  File has not been updated.
+#  Implicit rule search has been done.
+#  Last modified 2022-06-09 17:33:30
+#  File has been updated.
+#  Successfully updated.
+# variable set hash-table stats:
+# Load=0/32=0%, Rehash=0, Collisions=0/0=0%
 
 # files hash-table stats:
-# Load=12/1024=1%, Rehash=0, Collisions=2/33=6%
+# Load=13/1024=1%, Rehash=0, Collisions=2/35=6%
 # VPATH Search Paths
 
 # No `vpath' search paths.
@@ -299,6 +319,6 @@ Commands/Pass.o:
 # strcache size: total = 4096 / max = 4096 / min = 4096 / avg = 4096
 # strcache free: total = 4087 / max = 4087 / min = 4087 / avg = 4087
 
-# Finished Make data base on Thu Jun  9 16:54:09 2022
+# Finished Make data base on Thu Jun  9 17:34:00 2022
